move cachedbatch to separate project
