+++ debezium-server-cachedbatch-sinks/src/main/resources/META-INF/beans.xml
@@ -1,0 +1,8 @@
+<!--
+  ~ /*
+  ~  * Copyright memiiso Authors.
+  ~  *
+  ~  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+  ~  */
+  -->
+
