+++ debezium-server-cachedbatch-sinks/src/main/resources/conf/application.properties.example
@@ -1,0 +1,22 @@
+debezium.sink.type=iceberg
+debezium.sink.kinesis.region=eu-central-1
+debezium.source.connector.class=io.debezium.connector.postgresql.PostgresConnector
+debezium.source.offset.storage.file.filename=data/offsets.dat
+debezium.source.offset.flush.interval.ms=0
+debezium.source.database.hostname=localhost
+debezium.source.database.port=5432
+debezium.source.database.user=postgres
+debezium.source.database.password=postgres
+debezium.source.database.dbname=postgres
+debezium.source.database.server.name=tutorial
+debezium.source.schema.include.list=inventory
+
+# @TODO add defaults!!
+debezium.sink.iceberg.fs.defaultFS=xyz
+debezium.sink.batch.objectkey.mapper=xyz
+debezium.format.schemas.enable=xyz
+debezium.sink.iceberg.warehouse=xyz
+debezium.transforms=xyz
+
+
+# @TODO add multiple example!!
