+++ debezium-server-cachedbatch-sinks/src/test/java/io/debezium/server/cachedbatch/InfinispanCacheTestProfile.java
@@ -1,0 +1,27 @@
+/*
+ *
+ *  * Copyright memiiso Authors.
+ *  *
+ *  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ */
+
+package io.debezium.server.cachedbatch;
+
+import io.quarkus.test.junit.QuarkusTestProfile;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class InfinispanCacheTestProfile implements QuarkusTestProfile {
+
+  //This method allows us to override configuration properties.
+  @Override
+  public Map<String, String> getConfigOverrides() {
+    Map<String, String> config = new HashMap<>();
+
+    config.put("debezium.sink.batch.row-limit", "3");
+
+    return config;
+  }
+}
