+++ debezium-server-cachedbatch-sinks/src/test/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,0 +1,17 @@
+#
+# /*
+#  * Copyright memiiso Authors.
+#  *
+#  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+#  */
+#
+
+#
+# /*
+#  * Copyright memiiso Authors.
+#  *
+#  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+#  */
+#
+
+io.debezium.server.cachedbatch.ConfigSource
